drivers: fbdev: dpu_everest: add support for DSC slice concept
Used for setting partial width and height for LCD, calculation of dsc_dec_sw should be equal to xres divided by the dsc_slice_num.

If a slice width is non-multiple of 3, DECON and DSIM must be set up to the dummy size based on the rule. However, some code was implemented without considering this point.

Created with reference to cal_9820 and Exynos9820 DPU driver. Also update the panel drivers panel_drv to account for this when setting multiresolution.